Deduplicate file validation and UTF-8 checking utilities
Java: Extract validateFilePath() into ExecuTorchRuntime and replace
inline checks in Module, LlmModule, and TrainingModule. This also
removes the "!!" typo in TrainingModule error messages.

C++: Move utf8_check_validity() from anonymous namespaces in
jni_layer_llama.cpp and jni_layer_asr.cpp into the shared
jni_helper.h/.cpp.
